refactor(LX TDD Training): Split Unit Test
Split out unit tests into two files

ICTRAINING-29
